Move sqrt to math_utils + cleanup
