+++ felt/src/bigint_felt.rs
@@ -417,7 +417,6 @@ impl<'a, const PH: u128, const PL: u128> Pow<FeltBigInt<PH, PL>> for &'a FeltBig
     }
 }
 
-
 impl<const PH: u128, const PL: u128> SubAssign for FeltBigInt<PH, PL> {
     fn sub_assign(&mut self, rhs: Self) {
         *self = &*self - &rhs;
